+++ CHANGELOG.md
@@ -1,11 +1,38 @@
-# Omi for Visual Studio Code Change Log
+# Omi for Visual Studio Code Change Log (Develop)
+## Version *
 
+### New Features
+
+- null
+
+### Bug Fixes
+
+- null
+---
+## Version 1.0.5 February 27, 2019
+
+### New Features
+
+- Omi ecosystem `[All branches and versions]`
+- Welcome omi `[Welcome Screen]`
+- Snippet `[Code completion]`
+- Code hint `[Code hover prompt]`
+- Jump `[Jump function]`
+- Dependencies menu
+- Json analysis menu
+- File explorer
+- Webview test 
+- In development...
+
+### Bug Fixes
+
+- null
+---
 ## Version 1.0.0 February 16, 2019
 
 ### New Features
 
 - Add message "hello omi!"
-
 - Add tree View Guide
 
 ### Bug Fixes